+++ playwright.config.ts
@@ -15,20 +15,32 @@ export default defineConfig({
     fullyParallel: !process.env.CI,
     /* Fail the build on CI if you accidentally left test.only in the source code. */
     forbidOnly: !!process.env.CI,
+    workers: process.env.CI ? 1 : undefined,
     /* Retry once on CI, never locally */
     retries: process.env.CI ? 1 : 0,
-    /* Opt out of parallel tests on CI. */
-    workers: process.env.CI ? 1 : undefined,
     /* Reporter to use. See https://playwright.dev/docs/test-reporters */
-    reporter: 'html',
+    reporter: [
+        [
+            'html',
+            {
+                open: process.env.CI ? 'never' : 'always', // if on CI then "never" otherwise "always" show
+            },
+        ],
+        [
+            'list',
+            {
+                printSteps: process.env.CI, // if on CI, print the steps
+            },
+        ],
+    ],
     /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
     use: {
         /* Base URL to use in actions like `await page.goto('/')`. */
-        baseURL: 'http://localhost:4173',
-        /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
-        trace: 'on-first-retry',
+        baseURL: 'http://127.0.0.1:5002',
         viewport: { width: 1280, height: 720 },
         screenshot: 'only-on-failure',
+        /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
+        trace: 'on',
     },
 
     /* Configure projects for major browsers */
@@ -41,11 +53,15 @@ export default defineConfig({
 
     /* Remove any lingering authentication state before starting the tests */
     webServer: {
-        command: 'rm -rf playwright/.auth && npm run preview',
-        url: 'http://localhost:4173/',
+        name: 'Vite build preview',
+        command: '',
+        url: 'http://127.0.0.1:5002',
         timeout: 180000,
-        reuseExistingServer: !process.env.CI,
+        reuseExistingServer: true,
         stdout: 'pipe',
         stderr: 'pipe',
     },
+
+    /* Clean stuff up after tests */
+    globalTeardown: './tests/teardown.ts',
 });