+++ playwright/fixtures.ts
@@ -11,7 +11,7 @@ export const test = baseTest.extend<{}, { workerStorageState: string }>({
         async ({ browser }, use) => {
             // Use parallelIndex as a unique identifier for each worker.
             const id = test.info().parallelIndex;
-            const fileName = path.resolve('playwright', '.auth', `${id}.json`);
+            const fileName = path.resolve('playwright', `.auth/${id}.json`);
 
             if (fs.existsSync(fileName)) {
                 // Reuse existing authentication state if any.
@@ -20,9 +20,9 @@ export const test = baseTest.extend<{}, { workerStorageState: string }>({
             }
 
             // Important: make sure we authenticate in a clean environment by unsetting storage state.
-            // @ts-ignore It accepts undefined but types are wrong.
             const page = await browser.newPage({
-                baseURL: 'http://localhost:4173',
+                baseURL: 'http://127.0.0.1:5002',
+                storageState: { cookies: [], origins: [] },
             });
 
             // Acquire a unique account, for example create a new one.
@@ -37,7 +37,7 @@ export const test = baseTest.extend<{}, { workerStorageState: string }>({
             // Go to the join page.
             await page.goto('/join');
 
-            // Create a new account.
+            // Create a new account by filling out the form.
             await page.getByTestId('username-field').fill(account.username);
             await page.getByTestId('password-field').fill(account.password);
             await page
@@ -46,13 +46,19 @@ export const test = baseTest.extend<{}, { workerStorageState: string }>({
             await page.getByTestId('join-button').click();
 
             // Wait for the final redirect URL to ensure that the cookies are actually set.
-            await page.waitForURL('/profile');
+            await page.waitForURL('/profile', {
+                waitUntil: 'domcontentloaded',
+            });
 
-            // End of authentication steps. We need to explicitly ask Playwright to save the indexedDB data stored by Firebase.
+            // Ask Playwright to save the indexedDB data stored by Firebase.
             await page
                 .context()
                 .storageState({ path: fileName, indexedDB: true });
+
+            // Close the page.
             await page.close();
+
+            // Use the stored state.
             await use(fileName);
         },
         { scope: 'worker' },