+++ src/locale/UITexts.ts
@@ -777,6 +777,8 @@ type UITexts = {
                 replace: string;
                 /** Remove a locale */
                 remove: string;
+                /** Menu button label for "other languages" (landing page)*/
+                menu: string;
             };
         };
         /** The keyboard shortcut reference dialog */
@@ -899,3 +901,4 @@ type UITexts = {
 };
 
 export { type UITexts as default };
+