+++ tests/end2end/subtitle-contrast.ts
@@ -22,8 +22,8 @@ function printAccessibilityScanResults(axeBuilderScanResults: AxeResults) {
         });
     });
 
-    expect(violations.length).toEqual(0),
-        'See stderr Attachment in report for Errors!';
+    (expect(violations.length).toEqual(0),
+        'See stderr Attachment in report for Errors!');
 }
 
 test('Test Wordplay homepage subtitles for WCAG violations', async ({