+++ CHANGELOG.md
@@ -3,6 +3,22 @@
 We'll note all notable changes in this file, including bug fixes, enhancements, and all closed issues.
 Dates are in `YYYY-MM-DD` format and versions are in [semantic versioning](http://semver.org/) format.
 
+## 0.6.37
+
+### Fixed
+
+- Less flaky authenticated end2end tests.
+- No redirect to login page prior to auth registration.
+
+### Added
+
+- Merged #875: Cycling language chooser on landing page.
+- Added Open Graph metadata for previews.
+
+### Maintenance
+
+- Updated minor versions of Svelte, SvelteKit, vite.
+
 ## 0.6.36
 
 ### Fixed